Elective MLProject questions 3-4.1
I have completed questions 3 and 4.1 of the elective MLProject. From tomorrow onwards I will start focusing on and working on my elective project which is the deep learning elective project. Once this and the capstone projects are completed I will start finishing all of the elective projects one  at a time.